refactored to custom hook
